Separate notification generate from open logic
GenerateNotificationOpenIntent used to handle everything related to
Intents for notifications since it had to create the full reverse
Activity trampoline and attach it to the notification. Now the
notification generation logic simply needs to always open the SDK's
invisible click tracking Activity. When the notification is open the
runtime logic will take the push payload into consideration to
trampoline to the destination Activity.
